requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
frozenlist==1.4.0
gTTS==2.3.2
idna==3.4
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
paho-mqtt==1.6.1
PyAudio==0.2.13
pygame==2.5.0
python-dotenv==1.0.0
python-vlc==3.0.18122
pyttsx3==2.90
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
